mindmate-backend/requirements.txt

flask-cors==3.0.10
flasgger==0.9.5
pymongo==4.5.0
python-dotenv==1.0.0
boto3==1.28.0
gunicorn==20.1.0
Flask==2.2.5
flask-cors==3.0.10
